requirements.txt
asteroid-filterbanks >=0.4
einops >=0.6.0
pyannote.core >= 5.0.0
pyannote.database >= 5.0.1
rich >= 12.0.0
semver >= 3.0.0
soundfile >= 0.12.1
torch >= 2.0.0
torch_audiomentations >= 0.11.0
pytorch_lightning
numpy
matplotlib
functools
torchmetrics
itertools
